Moving cart to the right side
