+++ client/src/layouts/homeLayout.tsx
@@ -1,3 +1,4 @@
+import { Box } from '@chakra-ui/react';
 import React from 'react';
 import { Outlet } from 'react-router-dom';
 import 'react-toastify/dist/ReactToastify.css';
@@ -9,7 +10,9 @@ const HomeLayout: React.FC = () => {
     <main>
       <AuthProvider>
         <Navbar />
-        <Outlet />
+        <Box as={'section'} width="100%" height="90vh">
+          <Outlet />
+        </Box>
       </AuthProvider>
     </main>
   );
